added Hero section
